fix chz llm userid全部改为int64

+++ controller/github.go
@@ -146,7 +146,7 @@ func GitHubOAuth(c *gin.Context) {
 				inviterId, _ = model.GetUserIdByAffCode(affCode.(string))
 			}
 
-			if err := user.Insert(inviterId); err != nil {
+			if err := user.Insert(int64(inviterId)); err != nil {
 				c.JSON(http.StatusOK, gin.H{
 					"success": false,
 					"message": err.Error(),
@@ -199,7 +199,7 @@ func GitHubBind(c *gin.Context) {
 	session := sessions.Default(c)
 	id := session.Get("id")
 	// id := c.GetInt("id")  // critical bug!
-	user.Id = id.(int)
+	user.Id = int64(id.(int))
 	err = user.FillUserById()
 	if err != nil {
 		common.ApiError(c, err)
