+++ controller/group.go
@@ -24,7 +24,7 @@ func GetGroups(c *gin.Context) {
 func GetUserGroups(c *gin.Context) {
 	usableGroups := make(map[string]map[string]interface{})
 	userGroup := ""
-	userId := c.GetInt("id")
+	userId := c.GetInt64("id")
 	userGroup, _ = model.GetUserGroup(userId, false)
 	for groupName, ratio := range ratio_setting.GetGroupRatioCopy() {
 		// UserUsableGroups contains the groups that the user can use
